+++ grid2op/Action/BaseAction.py
@@ -468,6 +468,22 @@ class BaseAction(GridObjects):
         self._modif_curtailment = False
         self._modif_alarm = False
 
+    @classmethod
+    def process_shunt_data(cls):
+        if not cls.shunts_data_available:
+            # this is really important, otherwise things from grid2op base types will be affected
+            cls.attr_list_vect = copy.deepcopy(cls.attr_list_vect)
+            cls.attr_list_set = copy.deepcopy(cls.attr_list_set)
+            # remove the shunts from the list to vector
+            for el in ["shunt_p", "shunt_q", "shunt_bus"]:
+                if el in cls.attr_list_vect:
+                    try:
+                        cls.attr_list_vect.remove(el)
+                    except ValueError:
+                        pass
+            cls.attr_list_set = set(cls.attr_list_vect)
+        return super().process_shunt_data()
+    
     def copy(self) -> "BaseAction":
         # sometimes this method is used...
         return self.__deepcopy__()
@@ -524,6 +540,10 @@ class BaseAction(GridObjects):
 
         return res
 
+    @classmethod
+    def process_shunt_data(cls):
+        
+        return super().process_shunt_data()
     def __deepcopy__(self, memodict={}) -> "BaseAction":
         res = type(self)()