+++ grid2op/Backend/EducPandaPowerBackend.py
@@ -200,7 +200,7 @@ class EducPandaPowerBackend(Backend):
         self._compute_pos_big_topo()  # we highly recommend you to call this !
 
         # and now the thermal limit
-        self.thermal_limit_a = 1000 * np.concatenate(
+        self.thermal_limit_a = 1000. * np.concatenate(
             (
                 self._grid.line["max_i_ka"].values,
                 self._grid.trafo["sn_mva"].values
@@ -211,8 +211,8 @@ class EducPandaPowerBackend(Backend):
 
         # NB: this instance of backend is here for academic purpose only. For clarity, it does not handle
         # neither shunt nor storage unit.
-        self.shunts_data_available = False
-        self.set_no_storage()
+        type(self).shunts_data_available = False
+        type(self).set_no_storage()
 
     ###### modify the grid
     def apply_action(self, backendAction=None):
@@ -399,13 +399,13 @@ class EducPandaPowerBackend(Backend):
             )
             i += 1
 
-        # do not forget storage units !
-        i = 0
-        for bus_id in self._grid.storage["bus"].values:
-            res[self.storage_pos_topo_vect[i]] = (
-                1 if bus_id == self.storage_to_subid[i] else 2
-            )
-            i += 1
+        # # do not forget storage units !
+        # i = 0
+        # for bus_id in self._grid.storage["bus"].values:
+        #     res[self.storage_pos_topo_vect[i]] = (
+        #         1 if bus_id == self.storage_to_subid[i] else 2
+        #     )
+        #     i += 1
         return res
 
     def generators_info(self):
@@ -433,7 +433,7 @@ class EducPandaPowerBackend(Backend):
         ].values.astype(
             dt_float
         )  # in pu
-        load_v *= self._grid.res_bus.loc[self._grid.load["bus"].values][
+        load_v *= self._grid.bus.loc[self._grid.load["bus"].values][
             "vn_kv"
         ].values.astype(
             dt_float