+++ grid2op/Observation/baseObservation.py
@@ -851,7 +851,23 @@ class BaseObservation(GridObjects):
             }
 
         return res
-
+    
+    @classmethod
+    def process_shunt_data(cls):
+        if not cls.shunts_data_available:
+            # this is really important, otherwise things from grid2op base types will be affected
+            cls.attr_list_vect = copy.deepcopy(cls.attr_list_vect)
+            cls.attr_list_set = copy.deepcopy(cls.attr_list_set)
+            # remove the shunts from the list to vector
+            for el in ["_shunt_p", "_shunt_q", "_shunt_v", "_shunt_bus"]:
+                if el in cls.attr_list_vect:
+                    try:
+                        cls.attr_list_vect.remove(el)
+                    except ValueError:
+                        pass
+            cls.attr_list_set = set(cls.attr_list_vect)
+        return super().process_shunt_data()
+    
     @classmethod
     def process_grid2op_compat(cls):
         if cls.glop_version == cls.BEFORE_COMPAT_VERSION: