+++ grid2op/Reward/CombinedReward.py
@@ -41,8 +41,8 @@ class CombinedReward(BaseReward):
 
     """
 
-    def __init__(self):
-        BaseReward.__init__(self)
+    def __init__(self, logger=None):
+        BaseReward.__init__(self, logger=logger)
         self.reward_min = dt_float(0.0)
         self.reward_max = dt_float(0.0)
         self.rewards = {}