+++ grid2op/Reward/LinesCapacityReward.py
@@ -40,8 +40,8 @@ class LinesCapacityReward(BaseReward):
 
     """
 
-    def __init__(self):
-        BaseReward.__init__(self)
+    def __init__(self, logger=None):
+        BaseReward.__init__(self, logger=logger)
         self.reward_min = dt_float(0.0)
         self.reward_max = dt_float(1.0)