+++ CHANGELOG.rst
@@ -36,6 +36,8 @@ Change Log
 - [FIXED] a documentation issue https://github.com/rte-france/Grid2Op/issues/281
 - [FIXED] a bug preventing to use the `FromChronix2grid` chronics class when 
   there is an opponent on the grid.
+- [FIXED] a documentation issue https://github.com/rte-france/Grid2Op/issues/319
+  on notebook 11
 - [IMPROVED] adding the compatibility with logger in the reward functions.
 
 [1.7.0] - 2022-04-29