+++ CHANGELOG.rst
@@ -3,6 +3,8 @@ Change Log
 
 [TODO]
 --------------------
+- [???] do not require "copy" for the backend (make something in case we cannot copy it)
+- [???] allow some extra arguments in creation of backend in the env and the runner
 - [???] closer integration with `gym` especially the "register env", being able to 
   create an env from a string etc.
 - [???] clean the notebook on RL
@@ -30,6 +32,22 @@ Change Log
   it in parallel and transfer the resulting data.
 - [???] "asynch" multienv
 - [???] properly model interconnecting powerlines
+
+[1.7.1] - 2022-xx-yy
+-----------------------
+- [FIXED] a documentation issue https://github.com/rte-france/Grid2Op/issues/281
+- [FIXED] a bug preventing to use the `FromChronix2grid` chronics class when 
+  there is an opponent on the grid.
+- [FIXED] a documentation issue https://github.com/rte-france/Grid2Op/issues/319
+  on notebook 11
+- [FIXED] some issues when the backend does not support shunts data (caused during the
+  computation of the size of the observation) Tests are now performed in
+  `grid2op/tests/test_educpp_backend.py`
+- [IMPROVED] adding the compatibility with logger in the reward functions.
+- [IMPROVED] when there is a game over caused by redispatching, the observation is
+  not updated, as it is the case for other type of game over (improved consistency)
+- [IMPROVED] it is now possible to make an environment with a backend that
+  cannot be copied.
 
 [1.7.0] - 2022-04-29
 ---------------------